pyproject.toml
[tool.poetry]
name = "changeloginator"
version = "0.1.0"
description = "Generates a changelog from the last tag to the latest commit of the curent branch"
authors = ["Lonnie Souder II <lonnie.souder.ii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
GitPython = "3.1.18"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
